implemented changes to restart ffmpeg process

+++ GoDogSBPackage/GoDogSB.cs
@@ -12,6 +12,8 @@ namespace GoDogSBPackage
         private Thread bgWorker;
         private Process process;
 
+        public bool IsForcedStopped { get; set; }
+
         public GoDogSB()
         {
             logger = new GoDogLogger();
@@ -29,15 +31,20 @@ namespace GoDogSBPackage
 
         public void StartConversion()
         {
+            IsForcedStopped = false;
             bgWorker = new Thread(new ThreadStart(StartVideoConversion));
             bgWorker.Start();
         }
 
-        public void StopConversion()
+        public void StopConversion(bool forcedStop = false)
         {
             try
             {
-                process.Kill();
+                IsForcedStopped = forcedStop;
+                if (!process.HasExited)
+                {
+                    process.Kill();
+                }
                 bgWorker.Abort();
             }
             catch (Exception e)
@@ -55,7 +62,7 @@ namespace GoDogSBPackage
         {
             string inputFile = @"rtsp://admin:Foxhound1!@192.168.1.64/Streaming/Channels/1";
             string outputFile = @"rtmp://104.248.182.51/live/2";
-            string filArgs = string.Format("-rtsp_transport tcp -i \"{0}\" -c:v copy -c:a aac -b:a 128k -ar 44100 -f flv \"{1}\"", inputFile, outputFile);
+            string filArgs = string.Format("-stimeout 5000 -rtsp_transport tcp -i \"{0}\" -c:v copy -c:a aac -b:a 128k -ar 44100 -f flv \"{1}\"", inputFile, outputFile);
 
             process = new Process();
 
@@ -76,17 +83,26 @@ namespace GoDogSBPackage
                 {
                     this.Log(process.StandardError.ReadLine());
                 }
+
+                this.Log($"Responding: {process.Responding.ToString()}");
             }
             catch (Exception ex)
             {
-                this.Log("Convertion Failed");
-                this.Log(ex.ToString());
+                this.Log($"Exception: {ex.ToString()}");
             }
             finally
             {
                 process.WaitForExit();
                 process.Close();
-                this.Log("Convertion Completed");
+
+                if (!IsForcedStopped)
+                {
+                    this.Log("Restarting conversion process.");
+                    StopConversion();
+                    StartConversion();
+                }
+
+                this.Log("Conversion process exited.");
             }
         }
     }
